fix interceptor and moving to correct core folder
