+++ src/assets/app-settings.json
@@ -1,0 +1,3 @@
+{
+    "productsUrl":"http://localhost:3000/products"
+}
